LayoutTests/imported/w3c/web-platform-tests/workers/name-property-expected.txt

CONSOLE MESSAGE: line 20: ReferenceError: Can't find variable: SharedWorker

FAIL Test the name property of shared and dedicated workers via the name constructor option ReferenceError: Can't find variable: SharedWorker
PASS Declaring name as an accidental global must not cause a harness error for DedicatedWorkerGlobalScope 
FAIL name property value for DedicatedWorkerGlobalScope assert_true: property exists on the global expected true got false
PASS name property is replaceable for DedicatedWorkerGlobalScope 
